Add logging field for direct share selection
To help evaluate the new ranking algo we want to know if the package for
a direct share was also offered as one of the ranked apps.
Logging the position if it is part of the top apps, otherwise logging
-1.
Also added a test case I missed for HashedStringCache and removing the
test for hashing resulting in the same hashed string when called twice
from the ChooserActivityTest since we now cover that in the
HashedStringCacheTest.

Bug: 130658734
Test: Added more test cases in ChooserActivityTest and did manual
testing

Change-Id: I0e34a6bf64114d94197f62b8219652c33c03a410