Fix: Allow system IME to be disabled when another IME is available
In P, pre-installed system IME could be disabled if an ASCII non-aux IME
was available. With I1352bf42dcb9219a370a16a381ea39c43c57637a,
system-IMEs could no longer be disabled. This was unintentional as the
method name isValidSystemNonAuxAsciiCapableIme was incorrect and it
wasn't checking isSystem() earlier.

This change renames it to isValidNonAuxAsciiCapableIme and removes the
system check in order to revert to original P behavior.

Change-Id: If40de88ca0536a4446135047b6ccab64681675ec
Fix: 124536147
Test: atest InputMethodPreferenceTest
Test: atest InputMethodAndSubtypeUtilTest
Test: Manually verified using steps in bug.