Set activity state to PAUSED when making visible from STOPPING/STOPPED
An invisible activity state was STOPPED and was not updated while
making it visible. Once the activity became invisible again, the
activity won’t have to reenter the STOPPED state. But the window
surface was not destroyed and the window was in the surface destroy list
(WindowState.mDestroying was true).

In that case, when resuming the activity by finishing current top
activity (such as pressing back key), a transition animation was
set, but failed to play and had to wait until timeout.

Bug: 118349021
Bug: 132811175
Bug: 134986295
Test: ActivityRecordTests
Test: ActivityVisibilityTests
Test: ActivityLifecycleTests

Change-Id: I887740312de8f8580c58cde36e0f5137365b607c