Return null from  PackageManagerService.getInstallerPackageName for apex
When installing an apex, installerPackageName is not stored anywhere.
The main purpose of installerPackageName is to signal whether or not
the package can be uninstalled silently. If uninstaller is not same as
the original installer, we prompt the user for confirmation. Null means
"no installer" and therefore is a safe default.

Bug: 123920130
Test: manual testing
      $adb install -i com.android.shell foo.apex
      $adb reboot
      $adb shell pm list packages -i --apex-only
      observe that installer is listed as null for foo.apex
Change-Id: Ic4637a47aebf9fd5df07d2b1e481be12e2763e30