Hide lock icon while WAKE_AND_UNLOCK is happening.
The lock icon is on top of the front scrim, otherwise it wouldn't be visible on a
scrimmed bouncer (activities occluding the keyguard.)
Because of this, the black layer the covers most of the lock screen won't cover the
icon animation. We need to ensure that the icon is invisible while the biometric
unlock animation is running, and only unblock it after the keyguard has finished
fading away.

Test: manually unlock with fingerprint from AOD and LS
Test: press power while unlocking
Fixes: 132198859
Change-Id: Ied501586abe05dc333e4d76e4c045f32070595a7