Fixed screen decor layout frames
The initial frames for the screen decor were set to the dock frame. This
is incorrect because it incorporates the status bar, preventing
the frame from starting at the corners. Instead, set the frame to
unrestricted to allow frames in the status bar region.

This bug was introduced with the refactor ag/4525486

Fixes: 135706131
Test: No visible change on phone. Fixes issue with auto
Test: DisplayPolicyLayoutTests#testScreenDecorWindows
Change-Id: I6782a3585252df505b43d3e706b0332c609c7414