Add textview and framelayout inflation cases, refactor.
This change adds a base activity to inflate views, and then adds two
more views to inflate: one with 1000 TextViews and the other with 1000
FrameLayouts in a LinearLayout nested inside a ScrollView.

This change also adds launcher icons for each of the new cases for
ease in testing a launcher-based start.

Bug: 132721345
Test: Verified that the layout segments appear in traces as expected.

Change-Id: Id8c3e867a1ad5df5c3322ce5d69c9b4255042b69