Add proper color management into ColorFade.
ColorFade will take a screenshot into Surface and then use it to do animation
when the device is powering on and off. However, when EGLSurface is created, it
doesn't have proper color management attributes if the Surface is in Display
P3. For now wide color implies Display P3, and thus we hard code Display P3
here. For longer solution, we should lift this limit and make sure we manage
color properly.

BUG: b/136785946, b/131928312
Test: reproduced steps in bug no longer reproducible
Change-Id: I063244d8b850187e62c8c10c7712bf615fdcaf87