Only update overlays for target
When a target package would update, the OMS would update the stat of
overlays that target android as if the android overlays targeted the
target package. This would cause the overlay state to change to
STATE_MISSING_TARGET for android overlays.

This change will only update state for overlays that target the updated
package.

Bug: 137038000
Bug: 136549878
Test: manual debugging
Change-Id: I0430426d8e8588a0fb97096a1f2570a7930069b3
Merged-In: I0d545253179b3545307b88b3151d11fd9e6cbd5b